backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
redis>=4.1.0
python-dotenv>=0.19.0
pydantic>=1.10.7
pydantic-settings>=2.0.0